+++ changelog.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.3] - 2023-03-04
+BUG FIXES:
+- modified givenergy-modbus library to calulate inverter type from registers, not from serial number
+- Fixed day/Night rate stability in dynamic mode
+
 ## [2.1.3] - 2023-02-12
 
 BUG FIXES: